**Description of changes**
Vectorizing for loops, remove redundant code.

<!-- Please state what you've changed and how it might affect the user. -->

**Technical details**
Don't have large memory machine available atm for larger benchmark check.
For small 30k samples, `mean_absolute`, `x_ordered`, `correlation` become ~0.2 to 0.4 seconds instead of ~2 to 3 seconds, in overall 20seconds runtime. Sinkhorn dominating runtime.

<!-- Please state any technical details such as limitations, reasons for additional dependencies, benchmarks etc. here. -->

**Additional context**

<!-- Add any other context or screenshots here. -->