@eroell 2025-04-29
> I assume that the results are the same.

While we don't have tight CI tests here, verified with the snippet below that the two implementations generate the same result, by comparing the following four saved outputs from before and with this change, which pass a `np.allclose` check.

i.e. running this script once on main, once on current branch
```py
import numpy as np
import pertpy as pt
import scanpy as sc

adata = pt.dt.cinemaot_example()
sc.pp.sample(adata, n=1_000, replace=True, rng=42)
adata.X = adata.raw.X.copy()

sc.pp.pca(adata)

cot = pt.tl.Cinemaot()

de = cot.causaleffect(
    adata,
    pert_key="perturbation",
    control="No stimulation",
    return_matching=True,
    thres=0.34, # a threshold where for this dataset, the implementation of Xi matters
    smoothness=3e-5,
    eps=1e-3,
    solver="Sinkhorn",
    preweight_label="cell_type0528",
)

np.save("cinemaot_<branch_name>_cft.npy", adata.obsm["cf"])
np.save("cinemaot_<branch_name>_X.npy", de.X)
np.save("cinemaot_<branch_name>_embedding.npy", de.obsm["X_embedding"])
np.save("cinemaot_<branch_name>_ot.npy", de.obsm["ot"])
```