@eroell 2025-04-29
> Can you ott Jax jit the computation though? That should be the primary loss of performance.

It is. My attempts here on speeding it up have not been successful. The recommendation e.g. in their tutorial [here](https://ott-jax.readthedocs.io/en/latest/tutorials/basic_ot_between_datasets.html) does not give speedups when their LinearProblem has parameters `a` and `b`. It seems to give a slight speedup without `a`, `b` though, but this is just ca ~10% faster (across a range of sample sizes).